+++ src/test/java/tinqle/tinqleServer/domain/account/repository/AccountRepositoryTest.java
@@ -9,19 +9,14 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 import tinqle.tinqleServer.domain.account.model.Account;
-import tinqle.tinqleServer.domain.comment.model.Comment;
 import tinqle.tinqleServer.domain.comment.repository.CommentRepository;
 import tinqle.tinqleServer.domain.config.TestQueryDslConfig;
-import tinqle.tinqleServer.domain.feed.model.Feed;
 import tinqle.tinqleServer.domain.feed.repository.FeedRepository;
 
 import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static tinqle.tinqleServer.domain.account.template.AccountTemplate.*;
-import static tinqle.tinqleServer.domain.comment.template.CommentTemplate.createDummyChildCommentExceptId;
-import static tinqle.tinqleServer.domain.comment.template.CommentTemplate.createDummyParentCommentExceptId;
-import static tinqle.tinqleServer.domain.feed.template.FeedTemplate.createDummyFeedAExceptId;
 
 @DataJpaTest
 @ActiveProfiles({"test"})
@@ -101,85 +96,5 @@ public class AccountRepositoryTest {
 
         //then
         assertThat(exists).isTrue();
-    }
-
-    @Test
-    @DisplayName("피드 작성자가 댓글 작성시 자신을 제외한 댓글 단 사람 조회 - 성공")
-    public void findParentCommentAuthorByFeedDistinctExceptFeedAuthor_success() throws Exception {
-        //given
-        Account dummyAccountA = createDummyAccountA_ExceptId();
-        accountRepository.save(dummyAccountA);
-        Account dummyAccountB = createDummyAccountB_ExceptId();
-        accountRepository.save(dummyAccountB);
-        Feed feed = createDummyFeedAExceptId(dummyAccountA);
-        feedRepository.save(feed);
-        Comment comment = createDummyParentCommentExceptId(dummyAccountB, feed);
-        commentRepository.save(comment);
-        Comment feedAuthorComment = createDummyParentCommentExceptId(dummyAccountA, feed);
-        commentRepository.save(feedAuthorComment);
-
-        //when
-        List<Account> accounts = accountRepository.findCommentAuthorByFeedDistinctExceptFeedAuthor(feed, dummyAccountA);
-
-        //then
-        assertThat(accounts.size()).isEqualTo(1);
-    }
-
-    @Test
-    @DisplayName("피드 작성자가 댓글 작성시 자신을 제외한 댓글/대댓글 단 사람 조회 - 성공")
-    public void findAllCommentAuthorByFeedDistinctExceptFeedAuthor_success() throws Exception {
-        //given
-        Account dummyAccountA = createDummyAccountA_ExceptId();
-        accountRepository.save(dummyAccountA);
-        Account dummyAccountB = createDummyAccountB_ExceptId();
-        accountRepository.save(dummyAccountB);
-        Account dummyAccountC = createDummyAccountExceptId("닉네임");
-        accountRepository.save(dummyAccountC);
-        Feed feed = createDummyFeedAExceptId(dummyAccountA);
-        feedRepository.save(feed);
-        Comment comment = createDummyParentCommentExceptId(dummyAccountB, feed);
-        commentRepository.save(comment);
-        Comment feedAuthorComment = createDummyParentCommentExceptId(dummyAccountA, feed);
-        commentRepository.save(feedAuthorComment);
-        Comment childComment = createDummyChildCommentExceptId(dummyAccountC, comment);
-        commentRepository.save(childComment);
-        Comment childComment_B = createDummyChildCommentExceptId(dummyAccountC, comment);
-        commentRepository.save(childComment_B);
-
-        //when
-        List<Account> accounts = accountRepository.findCommentAuthorByFeedDistinctExceptFeedAuthor(feed, dummyAccountA);
-
-        //then
-        assertThat(accounts.size()).isEqualTo(2);
-    }
-
-    @Test
-    @DisplayName("피드 작성자, 부모 댓글 작성자 제외한 대댓글에 참여한 대댓글 작성자 조회 - 성공")
-    public void findChildCommentAuthorByParentCommentDistinctExceptAuthors_success() throws Exception {
-        //given
-        Account dummyAccountA = createDummyAccountA_ExceptId();
-        accountRepository.save(dummyAccountA);
-        Account dummyAccountB = createDummyAccountB_ExceptId();
-        accountRepository.save(dummyAccountB);
-        Account dummyAccountC = createDummyAccountExceptId("닉네임");
-        accountRepository.save(dummyAccountC);
-        Account dummyAccountD = createDummyAccountExceptId("닉네임2");
-        accountRepository.save(dummyAccountD);
-        Feed feed = createDummyFeedAExceptId(dummyAccountA);
-        feedRepository.save(feed);
-        Comment comment = createDummyParentCommentExceptId(dummyAccountB, feed);
-        commentRepository.save(comment);
-        Comment feedAuthorComment = createDummyParentCommentExceptId(dummyAccountA, feed);
-        commentRepository.save(feedAuthorComment);
-        Comment childComment = createDummyChildCommentExceptId(dummyAccountC, comment);
-        commentRepository.save(childComment);
-        Comment childComment_B = createDummyChildCommentExceptId(dummyAccountD, comment);
-        commentRepository.save(childComment_B);
-        //when
-        List<Account> accounts = accountRepository.findChildCommentAuthorByParentCommentDistinctExceptAuthors(comment, dummyAccountA, dummyAccountB, dummyAccountD);
-
-        //then
-        assertThat(accounts.size()).isEqualTo(1);
-        assertThat(accounts.get(0).getNickname()).isEqualTo(dummyAccountC.getNickname());
     }
 }
